role column. Delete disabled for administrators

+++ src/Views/admin/member/index.blade.php
@@ -31,6 +31,7 @@
                     <tr>                        
                         <td>{{ trans('panichd::admin.table-name') }}</td>
 						<td>{{ trans('panichd::admin.table-email') }}</td>
+						<td>{{ trans('panichd::admin.role') }}</td>
 						<td>{{ trans('panichd::admin.table-action') }}</td>
                     </tr>
                 </thead>
@@ -40,25 +41,37 @@
                         <td>{{ $member->name }}</td>
                         <td>{{ $member->email }}</td>
 						<td>
+						@if ($member->panichd_admin == '1')
+							<button type="button" disabled="disabled" class="btn btn-danger btn-xs">{{ trans('panichd::admin.admin') }}</button>
+						@elseif ($member->panichd_agent == '1')
+							<button type="button" disabled="disabled" class="btn btn-warning btn-xs">{{ trans('panichd::lang.agent') }}</button>
+						@else
+							<button type="button" disabled="disabled" class="btn btn-default btn-xs">{{ trans('panichd::lang.user') }}</button>
+						@endif
+						</td>
+						<td>
                             <button type="button" class="btn btn-default btn_member_modal" data-member_id="{{ $member->id }}" data-member_name="{{ $member->name }}" data-member_email="{{ $member->email }}" data-route="update" data-form_action="{{ route($setting->grab('admin_route').'.member.update', ['id' => $member->id ]) }}">{{ trans('panichd::admin.btn-edit') }}</button>
-							{!! link_to_route(
-							$setting->grab('admin_route').'.member.destroy', trans('panichd::admin.btn-delete'), $member->id,
-							[
-							'class' => 'btn btn-default deleteit',
-							'form' => "delete-$member->id",
-							"user" => $member->name
-							])
-                                !!}
-                            {!! CollectiveForm::open([
-                                            'method' => 'DELETE',
-                                            'route' => [
-                                                        $setting->grab('admin_route').'.member.destroy',
-                                                        $member->id
-                                                        ],
-                                            'id' => "delete-$member->id"
-                                            ])
-                            !!}
-                            {!! CollectiveForm::close() !!}
+							@if ($member->panichd_admin != '1')
+								{!! link_to_route(
+								$setting->grab('admin_route').'.member.destroy', trans('panichd::admin.btn-delete'), $member->id,
+								[
+								'class' => 'btn btn-default deleteit',
+								'form' => "delete-$member->id",
+								"user" => $member->name
+								])
+									!!}
+								
+								{!! CollectiveForm::open([
+												'method' => 'DELETE',
+												'route' => [
+															$setting->grab('admin_route').'.member.destroy',
+															$member->id
+															],
+												'id' => "delete-$member->id"
+												])
+								!!}
+								{!! CollectiveForm::close() !!}
+							@endif
                         </td>
                     </tr>
                 @endforeach
